Modifiez le type de données de nbre\_1 et nbre\_2 de int à float.
La fonction d'entrée pour nbre\_1 et nbre\_2 inclut désormais une chaîne f pour convertir l'entrée de l'utilisateur en flottant. Cela permet d'accepter et de calculer des valeurs décimales.

+++ calculatrice_v1.py
@@ -1,5 +1,5 @@
-nbre_1 = int(input("Entrez un nombre  : "))
-nbre_2 =  int(input("Entrez un nombre : "))
+nbre_1 = float(input("Entrez un nombre  : "))
+nbre_2 =  float(input("Entrez un nombre : "))
 
 # operateur = input("Entrez votre operatuer mathematique : ")
